test: cover responses file input normalization
